# public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/public/brands": {
            "get": {
                "summary": "Get all brands",
                "responses": {
                    "200": {
                        "description": "Brands fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Brand"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/brands": {
            "post": {
                "summary": "Create a new brand",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Brand created successfully"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update a brand",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a brand",
                "responses": {
                    "200": {
                        "description": "Brand deleted successfully"
                    }
                },
                        "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Brand": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {   
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "required": ["name", "slug", "user_id"]
            }
        }
    }       
}